[SYCL-MLIR] Add `-sycl-implementation` and `-sycl-target` options to `sycl` lowering
These options control what implementation to target (currently only
DPC++ is supported) and what target to lower to (currently only spir
target is supported).

Each SYCL implementation may support different targets.
