+++ mlir-sycl/lib/Conversion/SYCLToLLVM/CMakeLists.txt
@@ -1,5 +1,6 @@
 add_mlir_conversion_library(MLIRSYCLToLLVM
   DialectBuilder.cpp
+  DPCPP.cpp
   SYCLToLLVM.cpp
 
   ADDITIONAL_HEADER_DIRS
