+++ polygeist/tools/cgeist/Options.h
@@ -13,6 +13,7 @@
 #ifndef CGEIST_OPTIONS_H_
 #define CGEIST_OPTIONS_H_
 
+#include "mlir/Dialect/SYCL/IR/SYCLAttributes.h"
 #include "llvm/ADT/SmallString.h"
 #include "llvm/Passes/OptimizationLevel.h"
 #include "llvm/Support/CommandLine.h"
@@ -234,5 +235,10 @@ llvm::cl::opt<bool>
 static llvm::cl::opt<bool>
     SYCLDeviceOnly("sycl-device-only", llvm::cl::init(true),
                    llvm::cl::desc("Only emit device code in MLIR output"));
+
+static llvm::cl::opt<mlir::sycl::Implementation> SYCLImplementation(
+    llvm::cl::desc("use-opaque-pointers"),
+    llvm::cl::init(mlir::sycl::Implementation::DPCPP),
+    llvm::cl::values(clEnumVal(mlir::sycl::Implementation::DPCPP, "dpcpp")));
 
 #endif /* CGEIST_OPTIONS_H_ */
