+++ polygeist/tools/cgeist/driver.cc
@@ -612,10 +612,27 @@ static LogicalResult finalizeCUDA(mlir::PassManager &PM, Options &options) {
   return success();
 }
 
+static llvm::Expected<sycl::LoweringTarget>
+getSYCLTargetFromTriple(const llvm::Triple &Triple) {
+  switch (Triple.getArch()) {
+  case llvm::Triple::spir:
+    CGEIST_WARNING(llvm::WithColor::warning()
+                   << "Using the 32-bits spir target may lead to errors when "
+                      "lowering the `sycl` dialect.\n");
+    [[fallthrough]];
+  case llvm::Triple::spir64:
+    return sycl::LoweringTarget::SPIR;
+  default:
+    return llvm::createStringError(std::errc::not_supported,
+                                   "Cannot lower SYCL target \"%s\" to LLVM",
+                                   Triple.getTriple().c_str());
+  }
+}
+
 static LogicalResult finalize(mlir::MLIRContext &Ctx,
                               mlir::OwningOpRef<mlir::ModuleOp> &Module,
                               Options &options, llvm::DataLayout &DL,
-                              bool &LinkOMP) {
+                              const llvm::Triple &Triple, bool &LinkOMP) {
   mlir::PassManager PM(&Ctx);
   if (mlir::failed(enableOptionsPM(PM)))
     return failure();
@@ -701,6 +718,10 @@ static LogicalResult finalize(mlir::MLIRContext &Ctx,
       mlir::PassManager PM3(&Ctx);
       ConvertPolygeistToLLVMOptions Options;
       Options.dataLayout = DL.getStringRepresentation();
+      if (options.getCgeistOpts().getSYCLIsDevice()) {
+        Options.syclImplementation = SYCLImplementation;
+        Options.syclTarget = ExitOnErr(getSYCLTargetFromTriple(Triple));
+      }
       PM3.addPass(createConvertPolygeistToLLVM(Options));
       // PM3.addPass(mlir::createLowerFuncToLLVMPass(options));
       PM3.addPass(polygeist::createLegalizeForSPIRVPass());
@@ -762,7 +783,7 @@ createAndExecutePassPipeline(mlir::MLIRContext &Ctx,
   if (mlir::failed(optimizeCUDA(Ctx, Module, options)))
     return failure();
 
-  if (mlir::failed(finalize(Ctx, Module, options, DL, LinkOMP)))
+  if (mlir::failed(finalize(Ctx, Module, options, DL, Triple, LinkOMP)))
     return failure();
 
   return success();
